refactor: move grant/revoke_role_to_user to role_mgr
