+++ common/users/src/user_mgr.rs
@@ -143,32 +143,6 @@ impl UserApiProvider {
             .map_err(|e| e.add_message_back("(while revoke user privileges)"))
     }
 
-    pub async fn grant_role_to_user(
-        &self,
-        tenant: &str,
-        user: UserIdentity,
-        grant_role: String,
-    ) -> Result<Option<u64>> {
-        let client = self.get_user_api_client(tenant)?;
-        client
-            .grant_role(user, grant_role.clone(), None)
-            .await
-            .map_err(|e| e.add_message_back("(while grant role to user)"))
-    }
-
-    pub async fn revoke_role_from_user(
-        &self,
-        tenant: &str,
-        user: UserIdentity,
-        revoke_role: String,
-    ) -> Result<Option<u64>> {
-        let client = self.get_user_api_client(tenant)?;
-        client
-            .revoke_role(user, revoke_role.clone(), None)
-            .await
-            .map_err(|e| e.add_message_back("(while revoke role from user)"))
-    }
-
     // Drop a user by name and hostname.
     pub async fn drop_user(&self, tenant: &str, user: UserIdentity, if_exists: bool) -> Result<()> {
         let client = self.get_user_api_client(tenant)?;
